+++ apps/infrastructure/src/rds_aurora.tf
@@ -7,7 +7,7 @@ resource "aws_rds_cluster" "cms_database_cluster" {
   cluster_identifier     = "cms-database-cluster"
   engine                 = "aurora-postgresql"
   engine_mode            = "provisioned"
-  engine_version         = "14.6"
+  engine_version         = "14"
   database_name          = "strapidb"
   master_username        = "postgres"
   master_password        = aws_ssm_parameter.cms_database_password.value
@@ -22,9 +22,10 @@ resource "aws_rds_cluster" "cms_database_cluster" {
 }
 
 resource "aws_rds_cluster_instance" "cms_database_instance" {
-  cluster_identifier   = aws_rds_cluster.cms_database_cluster.id
-  instance_class       = "db.serverless"
-  engine               = aws_rds_cluster.cms_database_cluster.engine
-  engine_version       = aws_rds_cluster.cms_database_cluster.engine_version
-  db_subnet_group_name = aws_db_subnet_group.cms_db_subnet.name
+  cluster_identifier         = aws_rds_cluster.cms_database_cluster.id
+  instance_class             = "db.serverless"
+  engine                     = aws_rds_cluster.cms_database_cluster.engine
+  engine_version             = aws_rds_cluster.cms_database_cluster.engine_version
+  auto_minor_version_upgrade = true
+  db_subnet_group_name       = aws_db_subnet_group.cms_db_subnet.name
 }